fix clippy error and resolve comment
